Acción y esquema de borrado campaña actualizado

+++ src/actions/delete-campaign/index.js
@@ -6,6 +6,8 @@ import { createSafeAction } from "@/lib/create-safe-action";
 import { revalidatePath } from "next/cache";
 import { createAudtiLog } from "@/lib/create-audit-log";
 import { Action, EntityType } from "@prisma/client";
+import { FACEBOOK_API_GRAPH_URL } from "@/constants/facebook";
+import { getMetaAccessToken } from "@/lib/get-meta-access-token";
 
 export async function handler(data) {
   const { userId, orgId } = auth()
@@ -14,27 +16,45 @@ export async function handler(data) {
     error: "Unauthorized"
   }
 
-  const { id } = data
+  const { id, client_id } = data
 
   try {
-    const campaign = await prisma.campaign.delete({
-      where: {
-        id,
-        orgId
+
+    const access_token = await getMetaAccessToken(client_id)
+
+    const response = await fetch(`${FACEBOOK_API_GRAPH_URL}/${id}`, {
+      method: "DELETE",
+      headers: {
+        "Authorization": `OAuth ${access_token}`
       }
     })
 
-    await createAudtiLog({
-      action: Action.DELETE,
-      entityType: EntityType.CAMPAIGN,
-      entityTitle: campaign.name,
-      entityId: campaign.id
-    })
+    const data = await response.json()
 
-    revalidatePath(`/agency/${orgId}/campaigns`)
+    if (data.success) {
+      const campaign = await prisma.campaign.delete({
+        where: {
+          id,
+          orgId
+        }
+      })
 
-    return { ok: true, data: campaign }
+      await createAudtiLog({
+        action: Action.DELETE,
+        entityType: EntityType.CAMPAIGN,
+        entityTitle: campaign.name,
+        entityId: campaign.id
+      })
+
+      revalidatePath(`/agency/${orgId}/campaigns`)
+
+      return { ok: true, data: campaign }
+    } else {
+      throw new Error(data.error.message)
+    }
+
   } catch (error) {
+    console.log(error)
     return { error: "Error deleting the campaign!" }
   }
 }
